Adding a means of configuring the scheduler, currently implemented as settings sitting within a JSON object, JSON as it is very portable and can also be sent over another control socket if desired.
Adding the initialisation of multiple Schedulers, each initialising the correct number of input queues as specified by the config file.

Attempted to implement basic round robin, but currently blocks after the first packet.
It is noted that the framework is not realtime, but instead works of simulated time base (how do we even get performance of this, or are we thinking pure algorithm based performance increases)
Suggestion to handle the simulated time is to have all the schedulers get time tokens simultaneously to coordinate their outputs (maybe?)

TODO: Find a way to get or set or measure the mean packet length for WRR